+++ GameServerLib/Logic/Packets/PacketDefinitions/S2C/AttachMinimapIcon.cs
@@ -10,10 +10,10 @@ namespace LeagueSandbox.GameServer.Logic.Packets.PacketDefinitions.S2C
             : base(PacketCmd.PKT_S2_C_ATTACH_MINIMAP_ICON)
         {
             _buffer.Write(unit.NetId);
-            _buffer.Write((byte)unk1);
+            _buffer.Write(unk1);
             _buffer.Write(Encoding.Default.GetBytes(iconName)); // This is probably the icon name, but sometimes it's empty
             _buffer.Fill(0, 64 - iconName.Length);              // Example: "Quest"
-            _buffer.Write((byte)unk2);
+            _buffer.Write(unk2);
             _buffer.Write(Encoding.Default.GetBytes(unk3));
             _buffer.Fill(0, 64 - unk3.Length); // Example: "Recall"
             _buffer.Write(Encoding.Default.GetBytes(unk4));
