+++ GameServerLib/Logic/Packets/PacketDefinitions/S2C/AttentionPingResponse.cs
@@ -8,9 +8,9 @@ namespace LeagueSandbox.GameServer.Logic.Packets.PacketDefinitions.S2C
     {
         public AttentionPingResponse(ClientInfo player, AttentionPingRequest ping) : base(PacketCmd.PKT_S2_C_ATTENTION_PING)
         {
-            _buffer.Write((float)ping.X);
-            _buffer.Write((float)ping.Y);
-            _buffer.Write((int)ping.TargetNetId);
+            _buffer.Write(ping.X);
+            _buffer.Write(ping.Y);
+            _buffer.Write(ping.TargetNetId);
             _buffer.Write((int)player.Champion.NetId);
             _buffer.Write((byte)ping.Type);
             _buffer.Write((byte)0xFB); // 4.18
