+++ GameServerLib/Logic/Packets/PacketDefinitions/S2C/CreateMonsterCamp.cs
@@ -8,13 +8,13 @@ namespace LeagueSandbox.GameServer.Logic.Packets.PacketDefinitions.S2C
         public CreateMonsterCamp(float x, float y, float z, string iconName, byte campId, byte campUnk, float unk)
             : base(PacketCmd.PKT_S2_C_CREATE_MONSTER_CAMP)
         {
-            _buffer.Write((float)x);
-            _buffer.Write((float)z);
-            _buffer.Write((float)y);
+            _buffer.Write(x);
+            _buffer.Write(z);
+            _buffer.Write(y);
             _buffer.Write(Encoding.Default.GetBytes(iconName));
             _buffer.Fill(0, 64 - iconName.Length);
-            _buffer.Write((byte)campId);
-            _buffer.Write((byte)campUnk);
+            _buffer.Write(campId);
+            _buffer.Write(campUnk);
 
             /*buffer.Write((byte)0x64); // <-|
             buffer.Write((byte)0x15); //   |
@@ -22,7 +22,7 @@ namespace LeagueSandbox.GameServer.Logic.Packets.PacketDefinitions.S2C
             buffer.Write((byte)0x41); //   |
             buffer.Write((byte)0x0C); // <-|*/
             _buffer.Fill(0, 5);
-            _buffer.Write((float)unk);
+            _buffer.Write(unk);
         }
     }
 }
