+++ GameServerLib/Logic/Packets/PacketDefinitions/S2C/DestroyObject.cs
@@ -8,7 +8,7 @@ namespace LeagueSandbox.GameServer.Logic.Packets.PacketDefinitions.S2C
         public DestroyObject(AttackableUnit destroyer, AttackableUnit destroyed)
             : base(PacketCmd.PKT_S2_C_DESTROY_OBJECT, destroyer.NetId)
         {
-            _buffer.Write((uint)destroyed.NetId);
+            _buffer.Write(destroyed.NetId);
         }
     }
 }
