+++ GameServerLib/Logic/Packets/PacketDefinitions/S2C/FloatingText2.cs
@@ -9,7 +9,7 @@ namespace LeagueSandbox.GameServer.Logic.Packets.PacketDefinitions.S2C
         public FloatingText2(AttackableUnit u, string text, byte type, int unk) : base(PacketCmd.PKT_S2_C_FLOATING_TEXT, u.NetId)
         {
             _buffer.Fill(0, 10);
-            _buffer.Write((byte)type); // From 0x00 to 0x1B, 0x1C shows nothing and 0x1D bugsplats
+            _buffer.Write(type); // From 0x00 to 0x1B, 0x1C shows nothing and 0x1D bugsplats
             _buffer.Write((byte)0x00);
             _buffer.Write((byte)0x00);
             _buffer.Write((byte)0x00);
