+++ GameServerLib/Logic/Packets/PacketDefinitions/S2C/FogUpdate2.cs
@@ -12,19 +12,19 @@ namespace LeagueSandbox.GameServer.Logic.Packets.PacketDefinitions.S2C
             _buffer.Write((byte)0xFF);
             _buffer.Write((byte)0xFF);
             _buffer.Write((byte)0xFF);
-            _buffer.Write((int)0);
-            _buffer.Write((uint)unit.NetId); // Fog Attached, when unit dies it disappears
-            _buffer.Write((uint)idManager.GetNewNetId()); //Fog NetID
-            _buffer.Write((int)0);
-            _buffer.Write((float)unit.X);
-            _buffer.Write((float)unit.Y);
+            _buffer.Write(0);
+            _buffer.Write(unit.NetId); // Fog Attached, when unit dies it disappears
+            _buffer.Write(idManager.GetNewNetId()); //Fog NetID
+            _buffer.Write(0);
+            _buffer.Write(unit.X);
+            _buffer.Write(unit.Y);
             _buffer.Write((float)2500);
-            _buffer.Write((float)88.4f);
+            _buffer.Write(88.4f);
             _buffer.Write((float)130);
-            _buffer.Write((float)1.0f);
-            _buffer.Write((int)0);
+            _buffer.Write(1.0f);
+            _buffer.Write(0);
             _buffer.Write((byte)199);
-            _buffer.Write((float)unit.VisionRadius);
+            _buffer.Write(unit.VisionRadius);
         }
     }
 }
