+++ GameServerLib/Logic/Packets/PacketDefinitions/S2C/GameSpeed.cs
@@ -7,7 +7,7 @@ namespace LeagueSandbox.GameServer.Logic.Packets.PacketDefinitions.S2C
         public GameSpeed(float gameSpeed) 
             : base(PacketCmd.PKT_S2_C_GAME_SPEED)
         {
-            _buffer.Write((float)gameSpeed);
+            _buffer.Write(gameSpeed);
         }
     }
 }
