+++ GameServerLib/Logic/Packets/PacketDefinitions/S2C/HeroSpawn.cs
@@ -10,7 +10,7 @@ namespace LeagueSandbox.GameServer.Logic.Packets.PacketDefinitions.S2C
             : base(PacketCmd.PKT_S2_C_HERO_SPAWN)
         {
             _buffer.Write((int)player.Champion.NetId);
-            _buffer.Write((int)playerId); // player Id
+            _buffer.Write(playerId); // player Id
             _buffer.Write((byte)40); // netNodeID ?
             _buffer.Write((byte)0); // botSkillLevel Beginner=0 Intermediate=1
             if (player.Team == TeamId.TEAM_BLUE)
@@ -24,16 +24,16 @@ namespace LeagueSandbox.GameServer.Logic.Packets.PacketDefinitions.S2C
             _buffer.Write((byte)0); // isBot
             //buffer.Write((short)0; // botRank (deprecated as of 4.18)
             _buffer.Write((byte)0); // spawnPosIndex
-            _buffer.Write((int)player.SkinNo);
+            _buffer.Write(player.SkinNo);
             foreach (var b in Encoding.Default.GetBytes(player.Name))
-                _buffer.Write((byte)b);
+                _buffer.Write(b);
             _buffer.Fill(0, 128 - player.Name.Length);
             foreach (var b in Encoding.Default.GetBytes(player.Champion.Model))
-                _buffer.Write((byte)b);
+                _buffer.Write(b);
             _buffer.Fill(0, 40 - player.Champion.Model.Length);
-            _buffer.Write((float)0.0f); // deathDurationRemaining
-            _buffer.Write((float)0.0f); // timeSinceDeath
-            _buffer.Write((int)0); // UNK (4.18)
+            _buffer.Write(0.0f); // deathDurationRemaining
+            _buffer.Write(0.0f); // timeSinceDeath
+            _buffer.Write(0); // UNK (4.18)
             _buffer.Write((byte)0); // bitField
         }
     }
